feat: syntax highlighting, inline comments, diff scroll fix, file line select
- Fix diff scroll blocking: changed diff_scroll/description_scroll/file_tree_scroll from u16→usize; clamp on every scroll-down operation using diff_lines_cache.len()
- Syntax highlighting: new syntax.rs component using syntect (OnceLock-cached SyntaxSet/ThemeSet); used in diff_view, file_tree detail panel, and ReviewCompose context panel; diff lines show + with green bg, - with red bg, context with syntax colors
- Precompute diff_line_ext (file extension per diff line) on DiffLoaded for O(1) per-line ext lookup in render
- FileTree inline commenting: press → to focus detail panel; j/k navigate diff lines; selected line highlighted; c opens ReviewCompose with file:line pre-populated
- ReviewCompose shows code context panel (surrounding diff lines with syntax highlight) when coming from FileTree line selection
- ReviewCompose header shows file:line when composing inline comment; comment saved with file_path and line to draft
- SummaryPreview: Review Body placeholder explains how to generate it; Review Type selector has clearer title; keybind shows Enter/p → Submit to GitHub

+++ src/ui/screens/pr_detail.rs
@@ -149,7 +149,7 @@ fn render_description(frame: &mut Frame, app: &App, area: Rect, t: &Theme) {
     let para = Paragraph::new(md_lines)
         .block(block)
         .wrap(Wrap { trim: false })
-        .scroll((app.description_scroll, 0))
+        .scroll((app.description_scroll as u16, 0))
         .style(Style::default().fg(t.foreground).bg(t.background));
 
     frame.render_widget(para, area);
